requirements.txt
Flask
requests
scipy
pandas
Tinkerer
opencv-python
Pillow
keras
matplotlib
tensorflow



